refactor: #332 JsonDTO 분리
