+++ BBus/BBus/Global/DTO/BusPosByVehicleIdDTO.swift
@@ -7,30 +7,6 @@
 
 import Foundation
 
-struct JsonHeader: Codable {
-    let msgHeader: MessageHeader
-}
-
-struct JsonMessage<T: Codable>: Codable {
-    let msgHeader: MessageHeader
-    let msgBody: MessageBody<T>
-}
-
-struct MessageHeader: Codable {
-    let headerMessage, headerCD: String
-    let itemCount: Int
-
-    enum CodingKeys: String, CodingKey {
-        case headerMessage = "headerMsg"
-        case headerCD = "headerCd"
-        case itemCount
-    }
-}
-
-struct MessageBody<T: Codable>: Codable {
-    let itemList: [T]
-}
-
 struct BusPosByVehicleIdDTO: Codable {
 
     let stationOrd: String
